+++ attachments/src/main/java/com/backstage/app/attachment/configuration/properties/AttachmentProperties.java
@@ -23,6 +23,7 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.http.MediaType;
+import org.springframework.util.unit.DataSize;
 
 import java.time.Duration;
 import java.util.Set;
@@ -65,6 +66,8 @@ public class AttachmentProperties
 		private String secretKey;
 
 		private String bucket;
+
+		private DataSize uploadPartSize = DataSize.ofMegabytes(256);
 	}
 
 	@Getter